Tweaks to enable larger networks
Higher recursion limit for initial network generation. show_progress()
only updates the display every 100 pixels; on some machines
pygame.display.flip() seems to be a very expensive operation.

Node x and y position are now stored as uint16's. Later with the
firmware version that'll be changed back to uint8_t, presumably with
some way of changing between the two based on what compile is being
done.

+++ world.h
@@ -45,8 +45,8 @@ struct ant {
 //
 // All this can be stored in ROM.
 struct node {
-    uint8_t x;
-    uint8_t y;
+    uint16_t x;
+    uint16_t y;
 
     struct ant *ant;
 };
